apps/product_detail/serializers.py
from rest_framework import serializers
from .models import ProductDetail


class ProductDetailSerializer(serializers.ModelSerializer):
    class Meta:
        model = ProductDetail
        fields = '__all__'