requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==5.1.1
djangorestframework==3.15.2
jwt==1.3.1
mysqlclient==2.2.4
pycparser==2.22
PyMySQL==1.1.1
sqlparse==0.5.1
